requirements.txt
django==3.0.3
djangorestframework==3.11.0
pytest==7.4.0
pytest-django==4.5.2